Fix libvirt provisioning for recent Fedora
1) Because udev renames network devices from their original enpXsY to
eth0 (since they are virtual and not hardware), the interface given
by the ifcfg-* file does not exist when 'service network restart'
is run, and that command fails, which fails provisioning when openshift
is started.  This patch removes those files, but it also depends on
these vagrant commits to identify Fedora 21 and to ignore biosdevname:

https://github.com/mitchellh/vagrant/commit/b776fe74b88f6ac8deac6a91f4e886e205a2f3de
https://github.com/dcbw/vagrant/commit/760b9daab77c26df1a43af8f9c12254589d219d4

2) This patch moves the private network static IP configuration (10.245/16) to
eth1 and uses eth0 for the public network DHCP configuration.  This is done
because the libvirt DHCP server for eth1 provides addresses in the 10.245 space,
while the DHCP server for eth0 provides the expected 192.168.121/24
addresses that vagrant-libvirt reads.  If the default vagrant configuration
for eth0 is left, then eth1 receives a DHCP lease in the 10.245/16 space
and no interface receives an address in the 192.168.121/24 space.

3) NFS shared directories depend on the IP address of the guest, which
can change each time the vagrant machine is halted and restarted.  To
fix a bug where /etc/exports is badly written by vagrant-libvirt you need
this patch in vagrant-libvirt:

https://github.com/pradels/vagrant-libvirt/commit/d107b899c8a3d0d22a9f8f7803f1e5753e4872e4

4) make sure that your host firewall doesn't block NFS access.  You
might need to do something like:
    sudo firewall-cmd --permanent --add-service=nfs
    sudo firewall-cmd --permanent --add-service=rpc-bind
    sudo firewall-cmd --permanent --add-service=mountd
    sudo firewall-cmd --restart
See: http://nts.strzibny.name/vagrant-nfs-exports-on-fedora-21/

5) You might have to start the master first, or use --no-parallel.
Since the 'libvirt' provider allows parallel provisioning, the
minions will race with the master to access the certificates which
may not yet have been created by the master.