Always bring up openshift's desired network configuration on Vagrant provision
OpenShift's static configuration (10.245.x.x) will always be
in ifcfg-eth1, and that's the network configuration that should
be started.  But since the ifcfg configuration might have
been written after NetworkManager started, NM may have already
created a non-persistent "default DHCP" connection for the
interface, which is already active.  We want to make sure that
the openshift-specific configuration is used instead.