sdn: add multicast support
Multicast uses a common address space, so we can't match directly
on a destination address for each pod, meaning we have to have
flows with multiple actions.

This could cause scalability problems later, which might be able
to be fixed with OpenFlow groups and buckets.  Each VNID would
be a group, and inside the group each pod would get a bucket.
But that's for later.

In this PR, multicast from the VXLAN is sent directly to the
local delivery table (#120).  Local multicast traffic is first
sent to VXLAN tunnels (table #110) and then chained to local
delivery.